refactor: tidy up button components
