+++ components/controls/CloseButton.tsx
